add separate file for dynamic ips stuff
