+++ cmd/nodecmd/dynamic_ips.go
@@ -1,0 +1,95 @@
+// Copyright (C) 2022, Ava Labs, Inc. All rights reserved.
+// See the file LICENSE for licensing terms.
+package nodecmd
+
+import (
+	"context"
+
+	awsAPI "github.com/ava-labs/avalanche-cli/pkg/cloud/aws"
+	gcpAPI "github.com/ava-labs/avalanche-cli/pkg/cloud/gcp"
+
+	"github.com/ava-labs/avalanche-cli/pkg/constants"
+
+	"github.com/ava-labs/avalanche-cli/pkg/ansible"
+
+	"github.com/ava-labs/avalanche-cli/pkg/models"
+	"github.com/ava-labs/avalanche-cli/pkg/ux"
+)
+
+func getNodesWoEIPInAnsibleInventory(clusterNodes []string) []models.NodeConfig {
+	nodesWoEIP := []models.NodeConfig{}
+	for _, node := range clusterNodes {
+		nodeConfig, err := app.LoadClusterNodeConfig(node)
+		if err != nil {
+			continue
+		}
+		if nodeConfig.ElasticIP == "" {
+			nodesWoEIP = append(nodesWoEIP, nodeConfig)
+		}
+	}
+	return nodesWoEIP
+}
+
+func getPublicIPForNodesWoEIP(nodesWoEIP []models.NodeConfig) (map[string]string, error) {
+	lastRegion := ""
+	var ec2Svc *awsAPI.AwsCloud
+	var err error
+	publicIPMap := make(map[string]string)
+	var gcpCloud *gcpAPI.GcpCloud
+	ux.Logger.PrintToUser("Getting Public IPs for nodes without static IPs ...")
+	for _, node := range nodesWoEIP {
+		if lastRegion == "" || node.Region != lastRegion {
+			if node.CloudService == "" || node.CloudService == constants.AWSCloudService {
+				ec2Svc, err = awsAPI.NewAwsCloud(awsProfile, node.Region)
+				if err != nil {
+					return nil, err
+				}
+			}
+			lastRegion = node.Region
+		}
+		var publicIP map[string]string
+		var err error
+		if node.CloudService == constants.GCPCloudService {
+			if gcpCloud == nil {
+				gcpClient, projectName, _, err := getGCPCloudCredentials()
+				if err != nil {
+					return nil, err
+				}
+				gcpCloud, err = gcpAPI.NewGcpCloud(gcpClient, projectName, context.Background())
+				if err != nil {
+					return nil, err
+				}
+			}
+			publicIP, err = gcpCloud.GetInstancePublicIPs(node.Region, []string{node.NodeID})
+			if err != nil {
+				return nil, err
+			}
+		} else {
+			publicIP, err = ec2Svc.GetInstancePublicIPs([]string{node.NodeID})
+			if err != nil {
+				return nil, err
+			}
+		}
+		publicIPMap[node.NodeID] = publicIP[node.NodeID]
+	}
+	return publicIPMap, nil
+}
+
+func updateAnsiblePublicIPs(clusterName string) error {
+	clusterNodes, err := getClusterNodes(clusterName)
+	if err != nil {
+		return err
+	}
+	nodesWoEIP := getNodesWoEIPInAnsibleInventory(clusterNodes)
+	if len(nodesWoEIP) > 0 {
+		publicIP, err := getPublicIPForNodesWoEIP(nodesWoEIP)
+		if err != nil {
+			return err
+		}
+		err = ansible.UpdateInventoryHostPublicIP(app.GetAnsibleInventoryDirPath(clusterName), publicIP)
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
